requirements.txt
requests>=2.31.0
pymysql>=1.1.0
selenium>=4.15.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
python-dateutil>=2.8.2
elasticsearch>=8.0.0
pymilvus>=2.3.0
networkx>=3.0
scikit-learn>=1.3.0
faiss-gpu>=1.7.4
rank-bm25>=0.2.2
zhipuai>=2.0.0
sentence-transformers>=2.2.0
